# src/main/resources/static/openapi.yaml
openapi: 3.1.0
info:
  title: ForgeFlow API
  version: "0.3"
  description: |
    Describe an app in plain English and get a working, running app back.

    **Auth.** `POST /api/v1/auth/signup` or `/login` returns a JWT. Send it as
    `Authorization: Bearer <token>` on everything else.

    **Errors** are RFC 9457 problem details (`application/problem+json`) with
    a `traceId` you can quote when reporting a problem.

    | Status | Means |
    |---|---|
    | 401 | not signed in |
    | 403 | you can see this project, but your role can't do that |
    | 404 | doesn't exist - **or** you have no access to it (indistinguishable on purpose) |
    | 402 | a plan limit (`quota`, `limit`, `used`, `plan` in the body) |
    | 409 | conflicts with current state (a reply already running, nothing to retry, ...) |
    | 429 | rate limited - see `Retry-After` |

    Every response carries `traceparent` and `X-Trace-Id` headers. Send a W3C
    `traceparent` and ForgeFlow continues your trace.

    This document is checked against the running application by
    `OpenApiContractTest`: an endpoint added without being documented here
    fails the build.
servers:
  - url: /
security:
  - bearer: []
...
paths:
...
  /api/v1/projects/{projectId}/files/content:
    parameters: [{ $ref: "#/components/parameters/projectId" }]
    get:
      tags: [Files]
      summary: File content (READ)
      parameters: [{ name: path, in: query, required: true, schema: { type: string }, example: index.html }]
      responses:
        "200":
          description: The file
          content: { application/json: { schema: { type: object, properties: { path: { type: string }, content: { type: string } } } } }
        "404": { $ref: "#/components/responses/Problem" }
    put:
      tags: [Files]
      summary: Save a file by hand - create or overwrite (WRITE)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [path, content]
              properties:
                path: { type: string, example: js/cart.js }
                content: { type: string }
      responses:
        "200":
          description: Saved
          content: { application/json: { schema: { type: object, properties: { path: { type: string }, version: { type: integer }, sizeBytes: { type: integer } } } } }
        "400": { $ref: "#/components/responses/Problem" }
        "403": { $ref: "#/components/responses/Problem" }
        "413": { $ref: "#/components/responses/Problem" }
...
  /api/v1/projects/{projectId}/checkpoints/{checkpointId}/files/content:
    parameters:
      - { $ref: "#/components/parameters/projectId" }
      - { name: checkpointId, in: path, required: true, schema: { type: integer, format: int64 } }
    get:
      tags: [History]
      summary: One file as it was in this version (READ)
      parameters: [{ name: path, in: query, required: true, schema: { type: string } }]
      responses:
        "200": { description: The file, content: { application/json: { schema: { type: object, properties: { path: { type: string }, content: { type: string } } } } } }
        "404": { $ref: "#/components/responses/Problem" }
...
  /api/v1/projects/{projectId}/preview:
    parameters: [{ $ref: "#/components/parameters/projectId" }]
    get:
      tags: [Preview]
      summary: Get the running preview (READ)
      responses:
        "200": { description: Running, content: { application/json: { schema: { $ref: "#/components/schemas/Preview" } } } }
        "404": { description: Nothing running, content: { application/problem+json: { schema: { $ref: "#/components/schemas/Problem" } } } }
    post:
      tags: [Preview]
      summary: Start (or restart) a preview (WRITE)
      responses:
        "200": { description: Started, content: { application/json: { schema: { $ref: "#/components/schemas/Preview" } } } }
        "402": { $ref: "#/components/responses/QuotaExceeded" }
        "409": { description: No files to preview, content: { application/problem+json: { schema: { $ref: "#/components/schemas/Problem" } } } }
    delete:
      tags: [Preview]
      summary: Stop the preview (WRITE)
      responses:
        "204": { description: Stopped }
...
components:
  securitySchemes:
    bearer: { type: http, scheme: bearer, bearerFormat: JWT }
  parameters:
...
    projectId: { name: projectId, in: path, required: true, schema: { type: integer, format: int64 } }
...
  responses:
    Problem:
      description: Problem details
      content: { application/problem+json: { schema: { $ref: "#/components/schemas/Problem" } } }
    QuotaExceeded:
      description: A plan limit was reached
      content: { application/problem+json: { schema: { $ref: "#/components/schemas/QuotaProblem" } } }
...
  schemas:
    Problem:
      type: object
      properties:
        type: { type: string }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
        traceId: { type: string }
    QuotaProblem:
      allOf:
        - $ref: "#/components/schemas/Problem"
        - type: object
          properties:
            quota: { type: string, enum: [PROJECTS, PREVIEWS, AI_TOKENS_PER_DAY] }
            limit: { type: integer, format: int64 }
            used: { type: integer, format: int64 }
            plan: { type: string }
            upgrade: { type: string }
...
    Preview:
      type: object
      properties:
        url: { type: string }
        status: { type: string }
        startedAt: { type: string, format: date-time }
        expiresAt: { type: string, format: date-time }
...